requirements.txt
# Core Frameworks
langchain
langchain-community
langchain-huggingface
langchain-nvidia-ai-endpoints
langserve[all]
langchain_chroma
tiktoken
fastapi
uvicorn
pydantic

# Data & Vector DB
chromadb
pymupdf4llm
fastf1
pandas

# UI
gradio
streamlit
python-dotenv